fix: split MCP trajectory bounds/tests for clippy and readability ratchet
